backend/api/requirementes.txt
Django==2.0.4
django-cors-headers==2.2.0
djangorestframework==3.8.2
psycopg2==2.7.4
pytz==2018.4
